backend/pyproject.toml
[project]
name = "kyron-medical-backend"
version = "0.1.0"
description = "Kyron Medical Patient Scheduling API"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "openai>=1.50.0",
    "sentence-transformers>=3.3.0",
    "torch>=2.2.0",
    "numpy>=1.26.0",
    "redis>=5.0.0",
    "sendgrid>=6.11.0",
    "twilio>=9.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-multipart>=0.0.9",
    "websockets>=12.0",
    "deepgram-sdk>=3.0.0",
    "edge-tts>=6.1.0",
    "pydub>=0.25.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]
